refactor: move material UI components to their own folder
Moved the Material UI components into their own folder, taking the opportunity to rename the files
match the casing convention.
