+++ src/components/FileEntry.tsx
